src/daemon/watcher/requirements.txt
pika==1.2.0  # You can specify a version or remove the version constraint
psycopg2-binary==2.9.1  # or just psycopg2 if you prefer
